Update filenames and split up files
