+++ src/main.py
@@ -4,7 +4,7 @@ import arcade
 
 from arena.arena import Arena
 from arena.arena_loader import load_arena_by_name
-from collision import projectile_hits_player, projectile_hits_wall
+from collision import ordnance_hits_player, ordnance_hits_wall
 from constants import (
     ARENA,
     SCREEN_HEIGHT,
@@ -18,8 +18,8 @@ from debug_patrol_loop import DebugPatrolLoop
 from fullscreen import FullscreenController
 from global_input import GlobalInput, bind_global_inputs_to_keyboard
 from hud import Hud
+from ordnances.ordnance import update_ordnance
 from player_manager import PlayerManager
-from projectile import update_ordnance
 from sprite_lists import SpriteLists
 
 
@@ -84,8 +84,8 @@ class MyGame(arcade.Window):
             delta_time,
             self.sprite_lists,
         )
-        projectile_hits_wall(self.sprite_lists)
-        projectile_hits_player(delta_time, self.sprite_lists)
+        ordnance_hits_wall(self.sprite_lists)
+        ordnance_hits_player(delta_time, self.sprite_lists)
         self.hud.update()
 
     def on_draw(self):
